+++ storm-core/test/clj/org/apache/storm/supervisor_test.clj
@@ -23,20 +23,20 @@
   (:import [org.apache.storm.scheduler ISupervisor])
   (:import [org.apache.storm.utils Time Utils$UptimeComputer ConfigUtils])
   (:import [org.apache.storm.generated RebalanceOptions WorkerResources])
-  (:import [org.mockito Matchers Mockito])
+  (:import [org.apache.storm.testing.staticmocking MockedCluster])
   (:import [java.util UUID])
+  (:import [org.apache.storm Thrift])
+  (:import [org.mockito Mockito Matchers])
+  (:import [org.mockito.exceptions.base MockitoAssertionError])
   (:import [java.io File])
   (:import [java.nio.file Files])
-  (:import [org.apache.storm.utils Utils IPredicate]
-           [org.apache.storm.utils.staticmocking ConfigUtilsInstaller
-                                                 UtilsInstaller])
+  (:import [org.apache.storm.utils Utils IPredicate])
+  (:import [org.apache.storm.cluster StormClusterStateImpl ClusterStateContext ClusterUtils]
+           [org.apache.storm.utils.staticmocking ConfigUtilsInstaller UtilsInstaller])
   (:import [java.nio.file.attribute FileAttribute])
-  (:import [org.apache.storm Thrift])
-  (:import [org.apache.storm.utils Utils])
-  (:use [org.apache.storm config testing util log])
+  (:use [org.apache.storm config testing util log converter])
   (:use [org.apache.storm.daemon common])
-  (:require [org.apache.storm.daemon [worker :as worker] [supervisor :as supervisor]]
-            [org.apache.storm [cluster :as cluster]])
+  (:require [org.apache.storm.daemon [worker :as worker] [supervisor :as supervisor]])
   (:use [conjure core])
   (:require [clojure.java.io :as io]))
 
@@ -45,7 +45,7 @@
   [cluster supervisor-id port]
   (let [state (:storm-cluster-state cluster)
         slot-assigns (for [storm-id (.assignments state nil)]
-                        (let [executors (-> (.assignment-info state storm-id nil)
+                        (let [executors (-> (clojurify-assignment (.assignmentInfo state storm-id nil))
                                         :executor->node+port
                                         (Utils/reverseMap)
                                         clojurify-structure
@@ -248,7 +248,7 @@
       )))
 
 (defn get-heartbeat [cluster supervisor-id]
-  (.supervisor-info (:storm-cluster-state cluster) supervisor-id))
+  (clojurify-supervisor-info (.supervisorInfo (:storm-cluster-state cluster) supervisor-id)))
 
 (defn check-heartbeat [cluster supervisor-id within-secs]
   (let [hb (get-heartbeat cluster supervisor-id)
@@ -642,14 +642,13 @@
           fake-utils (proxy [Utils] []
                        (localHostnameImpl [] nil)
                        (makeUptimeComputer [] (proxy [Utils$UptimeComputer] []
-                                                (upTime [] 0))))]
+                                                (upTime [] 0))))
+          cluster-utils (Mockito/mock ClusterUtils)]
       (with-open [_ (ConfigUtilsInstaller. fake-cu)
-                  _ (UtilsInstaller. fake-utils)]
-        (stubbing [cluster/mk-storm-cluster-state nil]
+                  _ (UtilsInstaller. fake-utils)
+                  mocked-cluster (MockedCluster. cluster-utils)]
           (supervisor/supervisor-data auth-conf nil fake-isupervisor)
-          (verify-call-times-for cluster/mk-storm-cluster-state 1)
-          (verify-first-call-args-for-indices cluster/mk-storm-cluster-state [2]
-              expected-acls))))))
+          (.mkStormClusterStateImpl (Mockito/verify cluster-utils (Mockito/times 1)) (Mockito/any) (Mockito/eq expected-acls) (Mockito/any))))))
 
   (deftest test-write-log-metadata
     (testing "supervisor writes correct data to logs metadata file"
@@ -772,68 +771,68 @@
             childopts-with-ids (supervisor/substitute-childopts childopts worker-id topology-id port mem-onheap)]
         (is (= expected-childopts childopts-with-ids)))))
 
-(deftest test-retry-read-assignments
-  (with-simulated-time-local-cluster [cluster
-                                      :supervisors 0
-                                      :ports-per-supervisor 2
-                                      :daemon-conf {ConfigUtils/NIMBUS_DO_NOT_REASSIGN true
-                                                    NIMBUS-MONITOR-FREQ-SECS 10
-                                                    TOPOLOGY-MESSAGE-TIMEOUT-SECS 30
-                                                    TOPOLOGY-ACKER-EXECUTORS 0}]
-    (letlocals
-     (bind sup1 (add-supervisor cluster :id "sup1" :ports [1 2 3 4]))
-     (bind topology1 (Thrift/buildTopology
-                      {"1" (Thrift/prepareSpoutDetails
-                             (TestPlannerSpout. true) (Integer. 2))}
-                      {}))
-     (bind topology2 (Thrift/buildTopology
-                      {"1" (Thrift/prepareSpoutDetails
-                             (TestPlannerSpout. true) (Integer. 2))}
-                      {}))
-     (bind state (:storm-cluster-state cluster))
-     (bind changed (capture-changed-workers
-                    (submit-mocked-assignment
-                     (:nimbus cluster)
-                     (:storm-cluster-state cluster)
-                     "topology1"
-                     {TOPOLOGY-WORKERS 2}
-                     topology1
-                     {1 "1"
-                      2 "1"}
-                     {[1 1] ["sup1" 1]
-                      [2 2] ["sup1" 2]}
-                     {["sup1" 1] [0.0 0.0 0.0]
-                      ["sup1" 2] [0.0 0.0 0.0]
-                      })
-                    (submit-mocked-assignment
-                     (:nimbus cluster)
-                     (:storm-cluster-state cluster)
-                     "topology2"
-                     {TOPOLOGY-WORKERS 2}
-                     topology2
-                     {1 "1"
-                      2 "1"}
-                     {[1 1] ["sup1" 1]
-                      [2 2] ["sup1" 2]}
-                     {["sup1" 1] [0.0 0.0 0.0]
-                      ["sup1" 2] [0.0 0.0 0.0]
-                      })
-                    ;; Instead of sleeping until topology is scheduled, rebalance topology so mk-assignments is called.
-                    (.rebalance (:nimbus cluster) "topology1" (doto (RebalanceOptions.) (.set_wait_secs 0)))
-                    ))
-     (is (empty? (:launched changed)))
-     (bind options (RebalanceOptions.))
-     (.set_wait_secs options 0)
-     (bind changed (capture-changed-workers
-                    (.rebalance (:nimbus cluster) "topology2" options)
-                    (advance-cluster-time cluster 10)
-                    (heartbeat-workers cluster "sup1" [1 2 3 4])
-                    (advance-cluster-time cluster 10)
-                    ))
-     (validate-launched-once (:launched changed)
-                             {"sup1" [1 2]}
-                             (get-storm-id (:storm-cluster-state cluster) "topology1"))
-     (validate-launched-once (:launched changed)
-                             {"sup1" [3 4]}
-                             (get-storm-id (:storm-cluster-state cluster) "topology2"))
-     )))
+  (deftest test-retry-read-assignments
+    (with-simulated-time-local-cluster [cluster
+                                        :supervisors 0
+                                        :ports-per-supervisor 2
+                                        :daemon-conf {ConfigUtils/NIMBUS_DO_NOT_REASSIGN true
+                                                      NIMBUS-MONITOR-FREQ-SECS 10
+                                                      TOPOLOGY-MESSAGE-TIMEOUT-SECS 30
+                                                      TOPOLOGY-ACKER-EXECUTORS 0}]
+      (letlocals
+        (bind sup1 (add-supervisor cluster :id "sup1" :ports [1 2 3 4]))
+        (bind topology1 (Thrift/buildTopology
+                          {"1" (Thrift/prepareSpoutDetails
+                                 (TestPlannerSpout. true) (Integer. 2))}
+                          {}))
+        (bind topology2 (Thrift/buildTopology
+                          {"1" (Thrift/prepareSpoutDetails
+                                 (TestPlannerSpout. true) (Integer. 2))}
+                          {}))
+        (bind state (:storm-cluster-state cluster))
+        (bind changed (capture-changed-workers
+                        (submit-mocked-assignment
+                          (:nimbus cluster)
+                          (:storm-cluster-state cluster)
+                          "topology1"
+                          {TOPOLOGY-WORKERS 2}
+                          topology1
+                          {1 "1"
+                           2 "1"}
+                          {[1 1] ["sup1" 1]
+                           [2 2] ["sup1" 2]}
+                          {["sup1" 1] [0.0 0.0 0.0]
+                           ["sup1" 2] [0.0 0.0 0.0]
+                           })
+                        (submit-mocked-assignment
+                          (:nimbus cluster)
+                          (:storm-cluster-state cluster)
+                          "topology2"
+                          {TOPOLOGY-WORKERS 2}
+                          topology2
+                          {1 "1"
+                           2 "1"}
+                          {[1 1] ["sup1" 1]
+                           [2 2] ["sup1" 2]}
+                          {["sup1" 1] [0.0 0.0 0.0]
+                           ["sup1" 2] [0.0 0.0 0.0]
+                           })
+                        ;; Instead of sleeping until topology is scheduled, rebalance topology so mk-assignments is called.
+                        (.rebalance (:nimbus cluster) "topology1" (doto (RebalanceOptions.) (.set_wait_secs 0)))
+                        ))
+        (is (empty? (:launched changed)))
+        (bind options (RebalanceOptions.))
+        (.set_wait_secs options 0)
+        (bind changed (capture-changed-workers
+                        (.rebalance (:nimbus cluster) "topology2" options)
+                        (advance-cluster-time cluster 10)
+                        (heartbeat-workers cluster "sup1" [1 2 3 4])
+                        (advance-cluster-time cluster 10)
+                        ))
+        (validate-launched-once (:launched changed)
+          {"sup1" [1 2]}
+          (get-storm-id (:storm-cluster-state cluster) "topology1"))
+        (validate-launched-once (:launched changed)
+          {"sup1" [3 4]}
+          (get-storm-id (:storm-cluster-state cluster) "topology2"))
+        )))