+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 ## 2.0.0
+ * STORM-1571: Improvment Kafka Spout Time Metric
+ * STORM-1569: Allowing users to specify the nimbus thrift server queue size.
+ * STORM-1564: fix wrong package-info in org.apache.storm.utils.staticmocking
+ * STORM-1267: Port set_log_level
+ * STORM-1266: Port rebalance
+ * STORM-1265: Port monitor
+ * STORM-1572: throw NPE when parsing the command line arguments by CLI
+ * STORM-1273: port backtype.storm.cluster to java
  * STORM-1479: use a simple implemention for IntSerializer
  * STORM-1255: port storm_utils.clj to java and split Time tests into its
  * STORM-1566: Worker exits with error o.a.s.d.worker [ERROR] Error on initialization of server mk-worker